`Refactor code: remove unused hook, update dummy data, and modify page components`
