+++ app/(app)/cases/view/[id]/page.tsx
@@ -3,63 +3,8 @@ import React from "react";
 import { SingleCaseHeader } from "../_components/single-case-header";
 import { CaseOverview } from "../_components/case_overview";
 import { CaseUpdates } from "../_components/case-updates";
+import { demoData } from "@/lib/dummy-data";
 
-export const demoData = {
-  caseNumber: "CV/Wuse/233456789/2024",
-  title:
-    "Johnathan Smith and Associates vs. The Federal Housing Development Corporation",
-  filedDate: "12/02/2024",
-  claimant: {
-    name: "John Doe",
-    email: "johndoe@gmail.com",
-    address: "22 Maitama Close",
-    phone: "+23480123456",
-  },
-  court: {
-    division: "Kuje",
-    district: "Kuje District",
-    magistrate: "Adebayo Adekoya",
-  },
-  caseType: {
-    type: "Criminal",
-    kind: "Recovery Of Premise",
-    worth: "Less than a million",
-  },
-  filings: [
-    {
-      date: "12/02/2024",
-      documents: [
-        {
-          title: "",
-          items: [
-            { name: "Originating Applications", amount: 500.0 },
-            { name: "Motion Exparte", amount: 500.0 },
-          ],
-        },
-        {
-          title: "Exhibits",
-          items: [{ name: "Car Light Picture", amount: 500.0 }],
-        },
-      ],
-    },
-    {
-      date: "15/02/2024",
-      documents: [
-        {
-          title: "",
-          items: [
-            { name: "Originating Applications", amount: 500.0 },
-            { name: "Motion Exparte", amount: 500.0 },
-          ],
-        },
-        {
-          title: "Exhibits",
-          items: [],
-        },
-      ],
-    },
-  ],
-};
 
 export default function SingleCasePage({ params }: { params: { id: string } }) {
